neural-core/requirements.txt
crewai
typer
sqlmodel
sqlalchemy
python-dotenv
openai
